+++ adapters/Pangle/PangleAdapter/GADMediationAdapterPangleConstants.h
@@ -15,7 +15,7 @@
 #import <Foundation/Foundation.h>
 
 /// Pangle adapter version.
-static NSString *const GADMAdapterPangleVersion = @"7.9.1.1.1";
+static NSString *const GADMAdapterPangleVersion = @"7.9.1.1.3";
 
 /// Pangle adapter Pangle App ID parameter key.
 static NSString *const GADMAdapterPangleAppID = @"appid";